U4X-859: Externalise deployment of forms and metada from source code to external files
