+++ api/src/main/java/org/openmrs/module/ugandaemr/api/impl/UgandaEMRServiceImpl.java
@@ -4,46 +4,35 @@ import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.codehaus.jackson.map.ObjectMapper;
-import org.openmrs.CareSetting;
-import org.openmrs.Concept;
-import org.openmrs.ConceptNumeric;
-import org.openmrs.DrugOrder;
-import org.openmrs.Encounter;
-import org.openmrs.EncounterType;
-import org.openmrs.EncounterRole;
-import org.openmrs.Location;
-import org.openmrs.Obs;
-import org.openmrs.Order;
-import org.openmrs.OrderType;
-import org.openmrs.Patient;
-import org.openmrs.Provider;
-import org.openmrs.PatientIdentifier;
-import org.openmrs.PatientIdentifierType;
-import org.openmrs.PatientProgram;
-import org.openmrs.PatientProgramAttribute;
-import org.openmrs.ProgramAttributeType;
-import org.openmrs.Person;
-import org.openmrs.Relationship;
-import org.openmrs.TestOrder;
-import org.openmrs.Visit;
-import org.openmrs.User;
+import org.openmrs.*;
+import org.openmrs.api.*;
 import org.openmrs.api.context.Context;
-import org.openmrs.api.AdministrationService;
-import org.openmrs.api.EncounterService;
-import org.openmrs.api.VisitService;
-import org.openmrs.api.PersonService;
-import org.openmrs.api.PatientService;
-import org.openmrs.api.ConceptService;
-import org.openmrs.api.OrderService;
-import org.openmrs.api.APIException;
 import org.openmrs.api.impl.BaseOpenmrsService;
+import org.openmrs.module.Module;
+import org.openmrs.module.ModuleFactory;
+import org.openmrs.module.appframework.service.AppFrameworkService;
+import org.openmrs.module.dataexchange.DataImporter;
+import org.openmrs.module.emrapi.EmrApiConstants;
+import org.openmrs.module.emrapi.utils.MetadataUtil;
 import org.openmrs.module.htmlformentry.FormEntrySession;
+import org.openmrs.module.idgen.IdentifierSource;
+import org.openmrs.module.idgen.service.IdentifierSourceService;
+import org.openmrs.module.metadatadeploy.api.MetadataDeployService;
+import org.openmrs.module.metadatamapping.MetadataTermMapping;
+import org.openmrs.module.metadatamapping.api.MetadataMappingService;
 import org.openmrs.module.patientqueueing.api.PatientQueueingService;
 import org.openmrs.module.patientqueueing.mapper.PatientQueueMapper;
 import org.openmrs.module.patientqueueing.model.PatientQueue;
 import org.openmrs.module.stockmanagement.api.dto.DispenseRequest;
 import org.openmrs.module.ugandaemr.PublicHoliday;
 import org.openmrs.module.ugandaemr.UgandaEMRConstants;
+import org.openmrs.module.ugandaemr.activator.AppConfigurationInitializer;
+import org.openmrs.module.ugandaemr.activator.HtmlFormsInitializer;
+import org.openmrs.module.ugandaemr.activator.Initializer;
+import org.openmrs.module.ugandaemr.activator.JsonFormsInitializer;
+import org.openmrs.module.ugandaemr.api.deploy.bundle.CommonMetadataBundle;
+import org.openmrs.module.ugandaemr.api.deploy.bundle.UgandaAddressMetadataBundle;
+import org.openmrs.module.ugandaemr.api.deploy.bundle.UgandaEMRPatientFlagMetadataBundle;
 import org.openmrs.module.ugandaemr.api.queuemapper.Identifier;
 import org.openmrs.module.ugandaemr.api.queuemapper.PatientQueueVisitMapper;
 import org.openmrs.module.ugandaemr.api.UgandaEMRService;
@@ -59,6 +48,7 @@ import org.openmrs.module.ugandaemr.pharmacy.DispensingModelWrapper;
 import org.openmrs.module.ugandaemr.pharmacy.mapper.DrugOrderMapper;
 import org.openmrs.module.ugandaemr.pharmacy.mapper.PharmacyMapper;
 import org.openmrs.notification.Alert;
+import org.openmrs.notification.AlertService;
 import org.openmrs.order.OrderUtil;
 import org.openmrs.parameter.EncounterSearchCriteria;
 import org.openmrs.parameter.EncounterSearchCriteriaBuilder;
@@ -66,6 +56,7 @@ import org.openmrs.ui.framework.SimpleObject;
 import org.openmrs.util.OpenmrsUtil;
 import org.openmrs.module.stockmanagement.api.StockManagementService;
 
+import java.io.File;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.text.ParseException;
@@ -2102,5 +2093,454 @@ public class UgandaEMRServiceImpl extends BaseOpenmrsService implements UgandaEM
         if(specimenSource!=null) {
             Context.getAdministrationService().executeSQL(String.format(SPECIMEN_MANUAL_UPDATE_QUERY,specimenSource.getConceptId(), order.getOrderId()), false);
         }
+    }
+
+
+    public Map initaliseMetaData(){
+        AdministrationService administrationService = Context.getAdministrationService();
+        AppFrameworkService appFrameworkService = Context.getService(AppFrameworkService.class);
+        MetadataDeployService deployService = Context.getService(MetadataDeployService.class);
+        LocationService locationService = Context.getLocationService();
+
+        String metaDataFilePath=getMetadataPath("metadata");
+        Map results=new HashMap<>();
+
+        try {
+            // disable the reference app registration page
+            appFrameworkService.disableApp("referenceapplication.registrationapp.registerPatient");
+            // disable the start visit app since all data is retrospective
+            appFrameworkService.disableExtension("org.openmrs.module.coreapps.createVisit");
+            // the extension to the edit person details
+            appFrameworkService.disableExtension("org.openmrs.module.registrationapp.editPatientDemographics");
+
+            // disable apps on the Clinican facing dashboard added through coreapps 1.12.0
+            appFrameworkService.disableApp("coreapps.mostRecentVitals");
+            appFrameworkService.disableApp("coreapps.diagnoses");
+            appFrameworkService.disableApp("coreapps.latestObsForConceptList");
+            appFrameworkService.disableApp("coreapps.obsAcrossEncounters");
+            appFrameworkService.disableApp("coreapps.obsGraph");
+            appFrameworkService.enableApp("coreapps.visitByEncounterType");
+            appFrameworkService.disableApp("coreapps.dataIntegrityViolations");
+            appFrameworkService.disableApp("fingerprint.findPatient");
+            appFrameworkService.enableApp("ugandaemr.findPatient");
+            appFrameworkService.disableApp("ugandaemr.registrationapp.registerPatient");
+
+            // enable the relationships dashboard widget
+            appFrameworkService.enableApp("coreapps.relationships");
+
+            // Remove the BIRT reports app since it is no longer supported
+            appFrameworkService.disableApp("ugandaemr.referenceapplication.birtReports");
+
+            // Home page apps clean up
+            appFrameworkService.disableApp("referenceapplication.vitals"); // Capture Vitals
+            appFrameworkService.disableApp("coreapps.activeVisits"); // Active Visits
+
+            // form entry app on the home page
+            appFrameworkService.disableApp("xforms.formentry");
+            // disable the default find patient app to provide one which allows searching for patients at the footer of the search for patients page
+            appFrameworkService.disableApp("coreapps.findPatient");
+            // form entry extension in active visits
+            appFrameworkService.disableExtension("xforms.formentry.cfpd");
+
+            // install concepts
+            DataImporter dataImporter = Context.getRegisteredComponent("dataImporter", DataImporter.class);
+
+            log.info("Start import of Concepts");
+            importConcepts(dataImporter);
+
+            log.info("Start import of person attributes");
+            // TODO: Replace this with metadata deploy to be consistent with other person attribute types
+            dataImporter.importData(metaDataFilePath+"Person_Attribute_Types.xml");
+            log.info("Person Attributes imported");
+
+            log.info("Start import of UgandaEMR Privileges");
+            dataImporter.importData(metaDataFilePath+"Role_Privilege.xml");
+            log.info("UgandaEMR Privileges Imported");
+
+            log.info("Start import of UgandaEMR Visits");
+            dataImporter.importData(metaDataFilePath+"VisitTypes.xml");
+            log.info("UgandaEMR Visits Imported");
+
+            // install commonly used metadata
+            installCommonMetadata(deployService);
+
+
+            log.info("Start import of UgandaEMR Relationship Types");
+            dataImporter.importData(metaDataFilePath+"RelationshipTypes.xml");
+            log.info("UgandaEMR Relationship Types Imported");
+
+            log.info("Start import of Program related objects");
+            dataImporter.importData(metaDataFilePath+"Programs.xml");
+            log.info(" Program related objects Imported");
+
+            // run the initializers
+            for (Initializer initializer : getInitializers()) {
+                initializer.started();
+            }
+
+            // save defined global properties
+            administrationService.saveGlobalProperties(configureGlobalProperties());
+
+            // update the name of the default health center with that stored in the global property
+            Location healthCenter = locationService.getLocationByUuid("629d78e9-93e5-43b0-ad8a-48313fd99117");
+            healthCenter.setName(administrationService.getGlobalProperty(UgandaEMRConstants.GP_HEALTH_CENTER_NAME));
+            locationService.saveLocation(healthCenter);
+
+            String flagstatus = administrationService.getGlobalProperty("ugandaemr.patientflags.disabledFlags");
+
+            if (flagstatus != null) {
+                flagstatus=("'"+flagstatus.trim().replace(",","','")+"'").replace(",''","").replace("' ","'");
+                administrationService.executeSQL("update patientflags_flag set enabled=0 where name in (" + flagstatus.trim() + ")", false);
+            }
+
+            // cleanup liquibase change logs to enable installation of data integrity module
+            removeOldChangeLocksForDataIntegrityModule();
+
+            // generate OpenMRS ID for patients without the identifier
+            generateOpenMRSIdentifierForPatientsWithout();
+
+            //update concept name for concept id  163017 to ABC-3TC-LPV/r as fully specified
+            administrationService.executeSQL("UPDATE concept_name SET concept_name_type = 'FULLY_SPECIFIED',locale_preferred = 1 WHERE concept_name_id = 134334 and concept_id=163017", false);
+            administrationService.executeSQL("UPDATE concept_name SET locale_preferred = 0,concept_name_type = null WHERE concept_name_id = 134333 and concept_id=163017", false);
+
+            log.info("ugandaemr Module started");
+            results.put("status","success");
+            results.put("message","completed initialising metadata");
+            return results;
+
+
+        } catch (Exception e) {
+            Module mod = ModuleFactory.getModuleById("ugandaemr");
+            ModuleFactory.stopModule(mod);
+            results.put("status","failed");
+            results.put("message",e.getMessage());
+            return results;
+        }
+    }
+
+    private String  getMetadataPath(String type){
+        String appDataDir = OpenmrsUtil.getApplicationDataDirectory();
+        String path="";
+
+        if(type.equals("jsonforms")){
+           path =Context.getAdministrationService().getGlobalProperty("ugandaemr.metadata.jsonFormPath");
+        }else if(type.equals("htmlforms")) {
+            path = Context.getAdministrationService().getGlobalProperty("ugandaemr.metadata.htmlnFormPath");
+        }else if(type.equals("metadata")) {
+            path = Context.getAdministrationService().getGlobalProperty("ugandaemr.metadata.path");
+        }
+        return appDataDir+path;
+    }
+
+
+
+    private void importConcepts(DataImporter dataImporter){
+        String metaDataFilePath=getMetadataPath("metadata");
+        log.info("import  to Concept Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept.xml");
+        log.info("import to Concept Table  Successful");
+
+        log.info("import  to Concept Name Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept_Name.xml");
+        log.info("import to Concept Name Table  Successful");
+
+        log.info("import  to Concept_Description Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept_Description.xml");
+        log.info("import to Concept_Description Table  Successful");
+
+        log.info("import  to Concept_Numeric Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept_Numeric.xml");
+        log.info("import to Concept_Numeric Table  Successful");
+
+        log.info("import  to Concept_Answer Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept_Answer.xml");
+        log.info("import to Concept_Answer Table  Successful");
+
+        log.info("import  to Concept_Set Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept_Set.xml");
+        log.info("import to Concept_Set Table  Successful");
+
+        log.info("import  to Concept_Reference Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept_Reference.xml");
+        log.info("import to Concept_Reference Table  Successful");
+
+        log.info("import  of  Concept Modifications Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Concept_Modifications.xml");
+        log.info("import to Concept Modifications Table  Successful");
+
+        log.info("import  of  Drugs  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/Drug.xml");
+        log.info("import of Drugs  Successful");
+
+        log.info("import  of  Drugs  Starting");
+        dataImporter.importData(metaDataFilePath+"appointment.xml");
+        log.info("import of Drugs  Successful");
+
+        log.info("import  of  ICD 11 concepts  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/icd_11/icd_11_import_concept.xml");
+        log.info("import of ICD 11 concepts  Successful");
+
+        log.info("import  of  ICD 11 concept_name Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/icd_11/icd_11_import_concept_name.xml");
+        log.info("import of ICD 11 concept_name  Successful");
+
+        log.info("import  of  ICD 11 concept_reference Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/icd_11/icd_11_import_concept_reference.xml");
+        log.info("import of ICD 11 concept_reference  Successful");
+
+        log.info("import  of  ICD 11 concept_map Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/icd_11/icd_11_import_concept_map.xml");
+        log.info("import of ICD 11 concept_map  Successful");
+
+        log.info("import  of  ICD 11 cause_of_death_set Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/cause_of_death_set.xml");
+        log.info("import of ICD 11 cause_of_death_set  Successful");
+
+        log.info("Move Non ICD Coded Diagnosis");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/icd_11/move_non_icd11-10-to-msc.xml");
+        log.info("Move non coded ICD 11 Diagnosis");
+
+        log.info("import  to Concept Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/tools-2024/Concept.xml");
+        log.info("import to Concept Table  Successful");
+
+        log.info("import  to Concept Name Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/tools-2024/Concept_Name.xml");
+        log.info("import to Concept Name Table  Successful");
+
+        log.info("import  to Concept_Description Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/tools-2024/Concept_Description.xml");
+        log.info("import to Concept_Description Table  Successful");
+
+        log.info("import  to Concept_Numeric Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/tools-2024/Concept_Numeric.xml");
+        log.info("import to Concept_Numeric Table  Successful");
+
+        log.info("import  to Concept_Answer Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/tools-2024/Concept_Answer.xml");
+        log.info("import to Concept_Answer Table  Successful");
+
+        log.info("import  to Concept_Set Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/tools-2024/Concept_Set.xml");
+        log.info("import to Concept_Set Table  Successful");
+
+        log.info("import  to Concept_Reference Table  Starting");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/tools-2024/Concept_Reference.xml");
+        log.info("import to Concept_Reference Table  Successful");
+
+        log.info("Retire Meta data");
+        dataImporter.importData(metaDataFilePath+"concepts_and_drugs/retire_meta_data.xml");
+        log.info("Retiring of meta data is Successful");
+    }
+
+    /**
+     * Generate patientIdentifier for old OpenMRS Migration to the new
+     */
+
+    protected PatientIdentifier generatePatientIdentifier() {
+        IdentifierSourceService iss = Context.getService(IdentifierSourceService.class);
+        IdentifierSource idSource = iss.getIdentifierSource(1); // this is the default OpenMRS identifier source
+        PatientService patientService = Context.getPatientService();
+
+        UUID uuid = UUID.randomUUID();
+
+        PatientIdentifierType patientIdentifierType = patientService.getPatientIdentifierTypeByUuid("05a29f94-c0ed-11e2-94be-8c13b969e334");
+
+        PatientIdentifier pid = new PatientIdentifier();
+        pid.setIdentifierType(patientIdentifierType);
+        String identifier = iss.generateIdentifier(idSource, "New OpenMRS ID with CheckDigit");
+        pid.setIdentifier(identifier);
+        pid.setPreferred(true);
+        pid.setUuid(String.valueOf(uuid));
+
+        return pid;
+
+    }
+
+    /**
+     * Generate an OpenMRS ID for patients who do not have one due to a migration from an old OpenMRS ID to a new one which contains a check-digit
+     **/
+    private void generateOpenMRSIdentifierForPatientsWithout() {
+        PatientService patientService = Context.getPatientService();
+        AdministrationService as = Context.getAdministrationService();
+        AlertService alertService = Context.getAlertService();
+
+        List<List<Object>> patientIds = as.executeSQL("SELECT patient_id FROM patient_identifier WHERE patient_id NOT IN (SELECT patient_id FROM patient_identifier p INNER JOIN patient_identifier_type pt ON (p.identifier_type = pt.patient_identifier_type_id AND pt.uuid = '05a29f94-c0ed-11e2-94be-8c13b969e334'))", true);
+
+        if (patientIds.size() == 0) {
+            // no patients to process
+            return;
+        }
+        // get the identifier source copied from RegistrationCoreServiceImpl
+
+        for (List<Object> row : patientIds) {
+            Patient p = patientService.getPatient((Integer) row.get(0));
+            // Create new Patient Identifier
+            PatientIdentifier pid = generatePatientIdentifier();
+            pid.setPatient(p);
+            try {
+                log.info("Adding OpenMRS ID " + pid.getIdentifier() + " to patient with id " + p.getPatientId());
+                // Save the patient Identifier
+                patientService.savePatientIdentifier(pid);
+            } catch (Exception e) {
+                // log the error to the alert service but do not rethrow the exception since the module has to start
+                log.error("Error updating OpenMRS identifier for patient #" + p.getPatientId(), e);
+            }
+        }
+        log.info("All patients updated with new OpenMRS ID");
+    }
+
+    /**
+     * Configure the global properties for the expected functionality
+     *
+     * @return
+     */
+    private List<GlobalProperty> configureGlobalProperties() {
+        List<GlobalProperty> properties = new ArrayList<GlobalProperty>();
+        // The primary identifier type now uses metadata mapping instead of a global property
+        MetadataMappingService metadataMappingService = Context.getService(MetadataMappingService.class);
+        MetadataTermMapping primaryIdentifierTypeMapping = metadataMappingService.getMetadataTermMapping(EmrApiConstants.EMR_METADATA_SOURCE_NAME, EmrApiConstants.PRIMARY_IDENTIFIER_TYPE);
+        PatientIdentifierType openmrsIdType = Context.getPatientService().getPatientIdentifierTypeByUuid(PatientIdentifierTypes.NATIONAL_ID.uuid());
+
+        //overwrite if not set yet
+        if (!openmrsIdType.getUuid().equals(primaryIdentifierTypeMapping.getMetadataUuid())) {
+            primaryIdentifierTypeMapping.setMappedObject(openmrsIdType);
+            metadataMappingService.saveMetadataTermMapping(primaryIdentifierTypeMapping);
+        }
+
+        String ART_Patient_Number_Identifier = "";
+        // check if the ART patient number is to be displayed then add it here
+        if (Context.getAdministrationService().getGlobalProperty("ugandaemr.showARTPatientNumberIdentifier").equals("true")) {
+            log.info("Adding ART patient number to extra identifier types property");
+            ART_Patient_Number_Identifier = "," + PatientIdentifierTypes.ART_PATIENT_NUMBER.uuid();
+        }
+
+        String Research_Patient_Identifier = "";
+        // check if the ART patient number is to be displayed then add it here
+        if (Context.getAdministrationService().getGlobalProperty("ugandaemr.showResearchPatientIdentifier").equals("true")) {
+            log.info("Adding research patient number to extra identifier types property");
+            Research_Patient_Identifier = "," + PatientIdentifierTypes.RESEARCH_PATIENT_ID.uuid();
+        }
+
+        String Refugee_Identifier = "";
+        // check if the ART patient number is to be displayed then add it here
+        if (Context.getAdministrationService().getGlobalProperty("ugandaemr.showRefugeeIdentifier").equals("true")) {
+            log.info("Adding refugee identifier to extra identifier types property");
+            Refugee_Identifier = "," + PatientIdentifierTypes.REFUGEE_IDENTIFICATION_NUMBER.uuid();
+        }
+
+        // set the HIV care number and EID number as additional identifiers that can be searched for
+        properties.add(new GlobalProperty(EmrApiConstants.GP_EXTRA_PATIENT_IDENTIFIER_TYPES, PatientIdentifierTypes.HIV_CARE_NUMBER.uuid() + "," + PatientIdentifierTypes.EXPOSED_INFANT_NUMBER.uuid() + "," + PatientIdentifierTypes.IPD_NUMBER.uuid() + "," + PatientIdentifierTypes.ANC_NUMBER.uuid() + "," + PatientIdentifierTypes.PNC_NUMBER.uuid() + "," + ART_Patient_Number_Identifier + Research_Patient_Identifier + Refugee_Identifier));
+
+        // set the name of the application
+        properties.add(new GlobalProperty("application.name", "UgandaEMR - Uganda eHealth Solution"));
+
+        // Remove the regular expression to validate names
+        properties.add(new GlobalProperty("patient.nameValidationRegex", ""));
+
+        // the search mode for patients to enable searching any part of names rather than the beginning
+        properties.add(new GlobalProperty("patientSearch.matchMode", "ANYWHERE"));
+
+        // enable searching on parts of the patient identifier
+        // the prefix and suffix provide a % round the entered search term with a like
+        properties.add(new GlobalProperty("patient.identifierPrefix", "%"));
+        properties.add(new GlobalProperty("patient.identifierSuffix", "%"));
+
+        // the RegeX and Search patterns should be empty so that the prefix and suffix matching above can work
+        properties.add(new GlobalProperty("patient.identifierRegex", ""));
+        properties.add(new GlobalProperty("patient.identifierSearchPattern", ""));
+        // add telephone number and Marital status to the search widget
+        properties.add(new GlobalProperty("patient.listingAttributeTypes", "Telephone Number,Marital Status"));
+
+        // Form Entry Settings
+        properties.add(new GlobalProperty("FormEntry.enableDashboardTab", "true"));     // show as a tab on the patient dashboard
+        properties.add(new GlobalProperty("FormEntry.FormEntry.enableOnEncounterTab", "true"));
+
+        //HTML Form Entry Settings
+        properties.add(new GlobalProperty("htmlformentry.showDateFormat", "false"));//Disable date format display on form entry
+
+        //Birt Settings
+        properties.add(new GlobalProperty("birt.alwaysUseOpenmrsJdbcProperties", "false"));
+        properties.add(new GlobalProperty("birt.birtHome", OpenmrsUtil.getApplicationDataDirectory() + "birt" + File.separator + "birt-runtime-2_3_2" + File.separator + "ReportEngine"));
+        properties.add(new GlobalProperty("birt.datasetDir", OpenmrsUtil.getApplicationDataDirectory() + "birt" + File.separator + "reports" + File.separator + "datasets"));
+        properties.add(new GlobalProperty("birt.loggingDir", OpenmrsUtil.getApplicationDataDirectory() + "birt" + File.separator + "logs"));
+        properties.add(new GlobalProperty("birt.defaultReportDesignFile", "default.rptdesign"));
+        properties.add(new GlobalProperty("birt.loggingLevel", "OFF"));
+        properties.add(new GlobalProperty("birt.mandatory", "false"));
+        properties.add(new GlobalProperty("birt.outputDir", OpenmrsUtil.getApplicationDataDirectory() + "birt" + File.separator + "reports" + File.separator + "output"));
+        properties.add(new GlobalProperty("birt.reportDir", OpenmrsUtil.getApplicationDataDirectory() + "birt" + File.separator + "reports"));
+        properties.add(new GlobalProperty("birt.reportOutputFile", OpenmrsUtil.getApplicationDataDirectory() + "birt" + File.separator + "reports" + File.separator + "output" + File.separator + "ReportOutput.pdf"));
+        properties.add(new GlobalProperty("birt.reportOutputFormat", "pdf"));
+        properties.add(new GlobalProperty("birt.reportPreviewFile", OpenmrsUtil.getApplicationDataDirectory() + "birt" + File.separator + "reports" + File.separator + "output" + File.separator + "ReportPreview.pdf"));
+
+        // disable the appointmentshedulingui which currently has issues
+        properties.add(new GlobalProperty("appointmentschedulingui.started", "false"));
+
+        // Exclude temporary reporting tables by database backup module
+        properties.add(new GlobalProperty("databasebackup.tablesExcluded", "ugandaemr_105_eid,ugandaemr_106a1a"));
+
+        // the name of the custom registration app
+        properties.add(new GlobalProperty("registrationapp.customRegistrationAppId", "ugandaemr.registrationapp.registerPatient"));
+
+        // enable the register patient button to appear on the search widget
+        properties.add(new GlobalProperty("coreapps.showRegisterPatientOnSearchWidget", "true"));
+
+        // mapping for creating visits without encounters to the default facility visit type
+        properties.add(new GlobalProperty("emrapi.EmrApiVisitAssignmentHandler.encounterTypeToNewVisitTypeMap", "default:7b0f5697-27e3-40c4-8bae-f4049abfb4ed"));
+        return properties;
+    }
+
+    private void installCommonMetadata(MetadataDeployService deployService) {
+        try {
+            log.info("Installing standard metadata using the packages.xml file");
+            MetadataUtil.setupStandardMetadata(getClass().getClassLoader());
+            log.info("Standard metadata installed");
+
+            log.info("Installing metadata");
+            log.info("Installing commonly used metadata");
+            deployService.installBundle(Context.getRegisteredComponents(CommonMetadataBundle.class).get(0));
+            log.info("Finished installing commonly used metadata");
+            log.info("Installing address hierarchy");
+            deployService.installBundle(Context.getRegisteredComponents(UgandaAddressMetadataBundle.class).get(0));
+            log.info("Finished installing addresshierarchy");
+
+            // install concepts
+            log.info("Installing patient flags");
+            deployService.installBundle(Context.getRegisteredComponents(UgandaEMRPatientFlagMetadataBundle.class).get(0));
+            log.info("Finished installing patient flags");
+
+        } catch (Exception e) {
+            Module mod = ModuleFactory.getModuleById("ugandaemr");
+            ModuleFactory.stopModule(mod);
+            throw new RuntimeException("failed to install the common metadata ", e);
+        }
+    }
+
+    private void removeOldChangeLocksForDataIntegrityModule() {
+        String gpVal = Context.getAdministrationService().getGlobalProperty("dataintegrity.database_version");
+        // remove data integrity locks for an version below 4
+        // some gymnastics to get the major version number from semver like 2.5.3
+        if ((gpVal == null) || new Integer(gpVal.substring(0, gpVal.indexOf("."))).intValue() < 4) {
+            AdministrationService as = Context.getAdministrationService();
+            log.warn("Removing liquibase change log locks for previously installed data integrity instance");
+            as.executeSQL("delete from liquibasechangelog WHERE ID like 'dataintegrity%';", false);
+        }
+    }
+
+    private List<Initializer> getInitializers() {
+        String jsonFormsPath=getMetadataPath("jsonforms");
+        String initaliseJsonForms=Context.getAdministrationService().getGlobalProperty("ugandaemr.metadata.initaliseJsonForms");
+        String htmlFormsPath=getMetadataPath("htmlforms");
+
+        String initalisehtmlForms=Context.getAdministrationService().getGlobalProperty("ugandaemr.metadata.initalisehHtmlForms");
+        List<Initializer> l = new ArrayList<Initializer>();
+        l.add(new AppConfigurationInitializer());
+        if(initaliseJsonForms.equals("true")){
+            l.add(new JsonFormsInitializer(UgandaEMRConstants.MODULE_ID,jsonFormsPath));
+        }
+        if(initalisehtmlForms.equals("true")){
+            l.add(new HtmlFormsInitializer(UgandaEMRConstants.MODULE_ID,htmlFormsPath));
+        }
+        return l;
     }
 }
